Tests/LibWeb/Text/expected/input-event-data.txt

insertText on input: data="a", inputType="insertText"
insertText on textarea: data="b", inputType="insertText"
deleteContentBackward: data=null, inputType="deleteContentBackward"
deleteContentForward: data=null, inputType="deleteContentForward"
Enter in textarea: data=null, inputType="insertLineBreak"
